--- ShoppingCart/WebContent/WEB-INF/web.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app xmlns="http://xmlns.jcp.org/xml/ns/javaee"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee http://xmlns.jcp.org/xml/ns/javaee/web-app_3_1.xsd"
-         version="3.1">
-    <display-name>ShoppingCart Management Web Application</display-name>
-
-    <context-param>
-        <param-name>jdbcURL</param-name>
-        <param-value>jdbc:mysql://localhost:3310/shoppingCart</param-value>
-    </context-param>
-
-    <context-param>
-        <param-name>jdbcUsername</param-name>
-        <param-value>root</param-value>
-    </context-param>
-
-    <context-param>
-        <param-name>jdbcPassword</param-name>
-        <param-value>nodejs</param-value>
-    </context-param>
-
-    <servlet>
-        <servlet-name>ControllerServlet</servlet-name>
-        <servlet-class>main.java.Controller.ControllerServlet</servlet-class>
-    </servlet>
-
-
-    <servlet-mapping>
-        <servlet-name>ControllerServlet</servlet-name>
-        <url-pattern>/</url-pattern>
-    </servlet-mapping>
-
-</web-app>
