--- ShoppingCart/src/main/java/Dao/ServiceDao.java
@@ -1,15 +1,0 @@
-package main.java.Dao;
-
-import java.sql.Connection;
-import java.sql.SQLException;
-
-public class ServiceDao {
-
-    public void connect(String jdbcURL, String jdbcUsername, String jdbcPassword, Connection jdbcConnection ) throws SQLException {
-
-    }
-
-    public void disconnect(Connection jdbcConnection) throws SQLException {
-
-    }
-}
