+++ src/main/java/com/globant/dao/ServiceDao.java
@@ -1,0 +1,15 @@
+package com.globant.dao;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+
+public class ServiceDao {
+
+    public void connect(String jdbcURL, String jdbcUsername, String jdbcPassword, Connection jdbcConnection ) throws SQLException {
+
+    }
+
+    public void disconnect(Connection jdbcConnection) throws SQLException {
+
+    }
+}
