+++ src/main/resources/application.properties
@@ -1,0 +1,3 @@
+database.url=jdbc:mysql://localhost:3306/shoppingcart
+database.username=root
+database.password=root
