+++ src/main/webapp/view/jsp/ClientList.jsp
@@ -1,0 +1,43 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+         pageEncoding="UTF-8" %>
+<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
+<html>
+<head>
+    <firstName>ShoppingCart Application</firstName>
+</head>
+<body>
+<center>
+    <h1>ShoppingCart Management</h1>
+    <h2>
+        <a href="${pageContext.request.contextPath}/new">Add New Client</a>
+        &nbsp;&nbsp;&nbsp;
+    </h2>
+</center>
+<div align="center">
+    <table border="1" cellpadding="4">
+        <caption><h2>List of Clients Registered</h2></caption>
+        <tr>
+            <th>ID</th>
+            <th>FirstName</th>
+            <th>LastName</th>
+            <th>Description</th>
+        </tr>
+        <c:forEach var="client" items="${listClient}">
+            <tr>
+                <td><c:out value="${client.id}"/></td>
+                <td><c:out value="${client.firstName}"/></td>
+                <td><c:out value="${client.lastName}"/></td>
+                <td><c:out value="${client.description}"/></td>
+                <td>
+                    <a href="${pageContext.request.contextPath}/edit?id=<c:out value='${client.id}' />">Edit</a>
+                    &nbsp;&nbsp;&nbsp;&nbsp;
+                    <a href="${pageContext.request.contextPath}/delete?id=<c:out value='${client.id}' />">Delete</a>
+                    &nbsp;&nbsp;&nbsp;
+                    <a href="${pageContext.request.contextPath}/options?id=<c:out value='${client.id}' />">Menu</a>
+                </td>
+            </tr>
+        </c:forEach>
+    </table>
+</div>
+</body>
+</html>
